Reorganize collection tasks
[noissue]
