+++ pulp_ansible/app/tasks/collections/__init__.py
@@ -1,0 +1,2 @@
+from pulp_ansible.app.tasks.collections.sync import sync  # noqa
+from pulp_ansible.app.tasks.collections.upload import import_collection  # noqa
